--- examples/minter/agents/mod.rs
@@ -1,7 +1,0 @@
-use super::*;
-pub(crate) mod token_admin;
-pub(crate) mod token_requester;
-
-pub fn default_max_count() -> Option<u64> {
-    Some(5)
-}
